Split big document upload page file into folder and sub components
