+++ ui/client/index.js
@@ -29,7 +29,7 @@ import Terminal from './terminal';
 import theme from './theme';
 import ViewDatasets from './components/ViewDatasets';
 import ViewDocuments, { ParagraphListings } from './documents';
-import UploadDocument from './documents/UploadForm';
+import UploadDocument from './documents/upload';
 import ViewModels from './components/ViewModels';
 import ViewRuns from './components/ViewRuns';
 // import DatasetRegistration from './datasets/Register';
